chore: added searchbar as component
